Rename file protobuf EventV1 converter
