+++ packages/models/src/index.ts
@@ -14,20 +14,20 @@ export * from "./attribute/protobufConverter.js";
 export * from "./email/email.js";
 
 export * from "./eservice/eservice.js";
+export * from "./eservice/eserviceEvents.js";
 export * from "./eservice/eserviceReadModelAdapter.js";
 export * from "./eservice/protobufConverterFromV1.js";
 export * from "./eservice/protobufConverterFromV2.js";
+export * from "./eservice/protobufConverterToV1.js";
 export * from "./eservice/protobufConverterToV2.js";
-export * from "./eservice/eserviceEvents.js";
-export * from "./eservice/eserviceEventV1Converter.js";
 
 export * from "./institution/institution.js";
 
 export * from "./risk-analysis/riskAnalysis.js";
 
+export * from "./tenant/protobufConverter.js";
 export * from "./tenant/tenant.js";
 export * from "./tenant/tenantEvents.js";
-export * from "./tenant/protobufConverter.js";
 
 export * from "./user/user.js";
 
@@ -44,14 +44,14 @@ export * from "./brandedIds.js";
 export * from "./errors.js";
 
 //  Generated models
-export * from "./gen/v1/eservice/eservice.js";
-export * from "./gen/v1/eservice/events.js";
-export * from "./gen/v2/eservice/eservice.js";
-export * from "./gen/v2/eservice/events.js";
 export * from "./gen/v1/agreement/agreement.js";
 export * from "./gen/v1/agreement/events.js";
 export * from "./gen/v1/agreement/state.js";
-export * from "./gen/v1/tenant/tenant.js";
-export * from "./gen/v1/tenant/events.js";
 export * from "./gen/v1/attribute/attribute.js";
 export * from "./gen/v1/attribute/events.js";
+export * from "./gen/v1/eservice/eservice.js";
+export * from "./gen/v1/eservice/events.js";
+export * from "./gen/v1/tenant/events.js";
+export * from "./gen/v1/tenant/tenant.js";
+export * from "./gen/v2/eservice/eservice.js";
+export * from "./gen/v2/eservice/events.js";
